Crate liquery

Structs

Query
A parsed query ready for use with a Context and Queryable

QueryReturn
Result of a query

Enums

EvalError
Errors found while evaluating query

ParseError
Errors found in parsing query

Token
AST tokens for the query string

Traits

Context
Trait describing an object that contains a dictionary of functions

Queryable
Trait describing an object that can be queried for variables

Type Definitions

EvalFunc
The General function signature for formatting function calls

EvalResult

ParseResult